+++ httpx/__init__.py
@@ -1,7 +1,7 @@
 from .__version__ import __description__, __title__, __version__
 from ._api import *
 from ._auth import *
-from ._client import *
+from ._clients import *
 from ._config import *
 from ._content import *
 from ._exceptions import *
