+++ httpx/_main.py
@@ -15,7 +15,7 @@ import rich.progress
 import rich.syntax
 import rich.table
 
-from ._client import Client
+from ._clients import Client
 from ._exceptions import RequestError
 from ._models import Response
 from ._status_codes import codes
