+++ scripts/lint
@@ -10,3 +10,4 @@ set -x
 
 ${PREFIX}ruff --fix $SOURCE_FILES
 ${PREFIX}ruff format $SOURCE_FILES
+${PREFIX}python scripts/unasync.py
